Move utils/score content to utils/score-helper
