+++ resources/assets/lib/utils/score-helper.ts
@@ -1,7 +1,8 @@
 // Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the GNU Affero General Public License v3.0.
 // See the LICENCE file in the repository root for full licence text.
 
-import ScoreJson from 'interfaces/score-json';
+import GameMode from 'interfaces/game-mode';
+import ScoreJson, { ScoreStatisticsAttribute } from 'interfaces/score-json';
 import core from 'osu-core-singleton';
 
 export function canBeReported(score: ScoreJson) {
@@ -23,3 +24,38 @@ export function hasReplay(score: ScoreJson) {
 export function hasShow(score: ScoreJson): score is ScoreJson & { best_id: number } {
   return score.best_id != null;
 }
+
+interface AttributeData {
+  attribute: ScoreStatisticsAttribute;
+  label: string;
+}
+
+const labelMiss = osu.trans('beatmapsets.show.scoreboard.headers.miss');
+
+export const modeAttributesMap: Record<GameMode, AttributeData[]> = {
+  fruits: [
+    { attribute: 'count_300', label: 'fruits' },
+    { attribute: 'count_100', label: 'ticks' },
+    { attribute: 'count_katu', label: 'drp miss' },
+    { attribute: 'count_miss', label: labelMiss },
+  ],
+  mania: [
+    { attribute: 'count_geki', label: 'max' },
+    { attribute: 'count_300', label: '300' },
+    { attribute: 'count_katu', label: '200' },
+    { attribute: 'count_100', label: '100' },
+    { attribute: 'count_50', label: '50' },
+    { attribute: 'count_miss', label: labelMiss },
+  ],
+  osu: [
+    { attribute: 'count_300', label: '300' },
+    { attribute: 'count_100', label: '100' },
+    { attribute: 'count_50', label: '50' },
+    { attribute: 'count_miss', label: labelMiss },
+  ],
+  taiko: [
+    { attribute: 'count_300', label: 'great' },
+    { attribute: 'count_100', label: 'good' },
+    { attribute: 'count_miss', label: labelMiss },
+  ],
+};
